specs: drop "The" in categorization titles of Markdown documentation
I think it makes sense to keep the "The" in place for the actual page's
title, but let's drop it from the categorization header, to make it
easier to find stuff, as the "The" isn't helpful to that.

In particular as we sometimes do it this and sometimes the other way so
far, hence let's stick to one common rule.