uki: reorder parts, describe stub more
The stub is the first thing that is loaded, so let's mention the stub first, the
kernel and initrd second and third, and other resources later. This makes the
initial descripition also match the order used later in the detailed description
of sections.

The description didn't say what the stub *is* (only what it does), and
that it it has a bunch of sections.