specs: import the table of PCRs 0—7 from systemd-cryptenroll(8)
This table is informative anyway. If somebody wants to make any realistic use of
it, they will usually want a full list of PCRs, because policies bind to e.g.
PCR 7 too. So let's copy in the remaining part from systemd-cryptenroll(8).
In the preamble it is mentioned that the list here is not authoritative.

In the man page, the discussion of when various PCRs change was included
directly in the table. It's tough to squeeze the extra text into the table,
so I moved after the table.

I dropped "(cont.d)" for the first row. When the table is rendered, an empty
field is visually clearer than the text.

https://sembr.org/ is used for wrapping.