spec: caret (^) compares lower instead of higher in step 5
Looking at the reference implementation in "systemd-analyze compare-versions"
the code handling the caret character is the same as for "-" and ".", right
before and after:

        /* Handle '-', which separates version and release, e.g 123.4-3.1.fc33.x86_64 */
        if (*a == '-' || *b == '-') {
                /* The string prefixed with '-' is older (e.g., 123-9 vs 123.1-1) */
                r = CMP(*a != '-', *b != '-');
                if (r != 0)
                        return r;

                a++;
                b++;
        }

        /* Handle '^'. Used for patched release. */
        if (*a == '^' || *b == '^') {
                r = CMP(*a != '^', *b != '^');
                if (r != 0)
                        return r;

                a++;
                b++;
        }

        /* Handle '.'. Used for point releases. */
        if (*a == '.' || *b == '.') {
                r = CMP(*a != '.', *b != '.');
                if (r != 0)
                        return r;

                a++;
                b++;
        }

so the caret character must be treated the same in the spec as well. This is not
a bug in the reference implementation, since when the sign of the comparison is
inverted, i.e. caret compares higher than not-caret, the comparison

        123^post1 < 123.a-1

would *not* hold true anymore, since "^" would win against "." at step 5. This
comparison must hold true, to be compatible with rpmdev-vercmp

Note: rpmdev-vercmp does compare 123^post1 < 123.a-1, but compares ^ > ., which
      systemd-analyze compares ^ < .