specs: implement standalone config files specification
Merging these patterns into XDG did not work out, so let's define the
rules independently for stateless and hermetic-usr configs.

The purpose of this document is to give answers to developers on what
the basic patterns need to be in order to support hermetic-usr and
stateless systems. First and foremost the recommendation is to use
libeconf wherever possible, as that already implements what's necessary
and we know it works. If that is not possible, then basic rules over
precedence between /etc/ and /usr/ are defined, along with drop-ins and
masking.

Defining all the low-level nuances as implemented by projects like
systemd for their specific use cases is not a goal, as these details
are intentionally left for the implementations. Defining what end users
need to do or expect is also not a goal, as each application's
documentation should be consulted instead.