Reconcile the handling of /boot and /efi in BLS and DPS
Fixes https://github.com/uapi-group/specifications/issues/35.

https://github.com/systemd/systemd/commit/6a488fa7cce8124fa885adf8a2f31363fe62f636
updated systemd-gpt-auto-generator as the first step. Now this commit
updates both specifications to match the implementation. The implementation
is actually much more complex than the specs, because it needs to take into
account a state where the mount points are missing, or used by something else,
or contain files. The specs just describe the desired end state:

  It is recommended to mount $BOOT (either XBOOTLDR or the ESP) to /boot/.
  If both XBOOTLDR and the ESP are present, the ESP should be mounted to /efi/.

  Effectively, this means that /boot/ is the location where new boot entries
  shall be written to. For the boot loader itself, the situation is more
  complicated, because its files may be located under /boot/ or /efi/, depending
  on whether XBOOTLDR exists.

So now the implementation, the man page, and the two specs should be in sync.