README: list all specifications
The README is prominently visible in two places: on the final website, and also
in the github repo. But on github there is no "menu on the left". So let's just
include the full list, which makes it easier to navigate (in both places).

Since the list includes short descriptions, the descriptions in the target pages
are also adjusted a bit to match the list.