specs/uki: say that only the .linux section is required, mention addons
People requested clarification as to what sections are "required" in a UKI. Say
that if .linux is present, it's a UKI. This is intentionally vague: depending on
the use of the UKI, some other sections will almost always be present, but also
depending on the intended use, those might be different sections. So let's keep
the specification simple and only define the minimum necessary.

When we have a proper specificiation for addons, we shall link it here.
(Or maybe we'll add it here?)

Closes #74.