specs/config-files: say that drop-ins are not always required
Be more explicit about the sort order: it is lexicographic, based on the file
name without path. Make the example more explicit.

Also drop the sentence that "drop-ins are not full replacements" — in practice
they are and I'm not sure what this sentence was trying to achieve.

Also, don't say "must parse all files" — in particular backup files and other
"stray" files should be ignored.

Resolves #48.