Make UKI description in boot_loader_specification.md more accurate and self-contained
The UKI description mentioned that it includes initrd and kernel command line resources and refers to the external systemd-stub reference for details. However, as described in the internal UKI reference, those are optional. This change reformulates the text such that it is made clear that UKIs should include an EFI stub loader and a kernel image and can include other, optional resources. Furthermore, the internal UKI reference is linked instead of the external systemd-stub reference.