bls: say that values are literal and update the initrd description
There was a discussion on fedora-devel whether grub2's addition of
interpolation of variables is in agreement with the spec. Let's make
it clearer that the value is just a value that should be used as
described.

For 'linux', do not say that the distro is supposed to create a single
directory for all kernels. It's supposed to be a single directory
per kernel version…

Also extend the description of 'initrd' to have the same format as
'linux'. One could think that they're somehow different. Extend the
example to show how two initrds can be combined.