bls: describe optional linux-appendroot key
The linux-appendroot key is an extension[1] to the bootloader
specification implemented by barebox since 2015[2].

When set, it instructs barebox to generate a suitable root=
command-line:

  - If global.bootm.root_dev is set to a partition known to barebox,
    barebox will detect the file system on it and generate the suitable
    root command-line argument

  - If global.bootm.root_dev is not set, barebox will use the file
    system containing the bootloader specification entries as the
    Linux rootfs.

The former is useful in A/B setups, where bootloader spec entries are
shipped as part of an immutable partition image, which is at odds with
customizing the root= option according to whether the A or B system is
used.

The latter is especially useful for remote file systems like NFS or 9P:
By placing bootloader specification entries directly into the rootfs,
these file systems can be booted remotely by just pointing barebox at
their remote path.

Describe this key in the specification with the same boolean syntax that
loader.conf is using for its boolean keys.

[1]: End of https://www.barebox.org/doc/latest/user/booting-linux.html#boot-loader-specification
[2]: https://github.com/barebox/barebox/commit/716fdbf18ca12feb81e26df729a5b8969e394e96

Signed-off-by: Ahmad Fatoum <ahmad@a3f.at>