Rename new page to "Linux File System Hierarchy"
I think we want to give this a unique name so that people can refer to it
easily. When linking to the (old) man page in a patch for Fedora Packaging Guidlines
I wrote "systemds file-hierarchy.7 page", because there was no obvious name.
With the new title we have LFSH as a nice abbreviation.

Also "file system", not "filesystem". We're specifically talking about the
whole "system of files", composed of multiple mount points. (I think using
"filesystem" for a single mount point makes a lot of sense. But for the whole
hierarchy, two words seem better.)