LFSH: update text after conversion from man page
Obviously, this is not a man page anymore, so some things needed to be reworded.
But also it's not part of the systemd project anymore, so we need to make the
text a bit more generic too.

Other issues noticed in the initial review when moving the page are fixed.