LFSH: tone down text about shared ownership in /usr/share
While it is true that shared ownership exists, there is no clear difference
between /usr/share and /usr/lib in this regard. In particular, systemd itself
makes heavy use of /usr/lib/sysctl.d and /usr/lib/tmpfiles.d and various other
directories which are great examples of shared ownership.
So tone this down to only remind the reader that the concept exists.