LFSH: describe /usr/include/
This is an imporant directory. It is not used at runtime, so the old man
page didn't need to describe it when it was part of the systemd project, but
since now we're turning this into a standalone document, we should cover it
to be comprehensive.

Both hier and FHS mention the page but only provide a very short description,
even shorter than the one added in this patch.