LFSH: say that libexec is not recommended
libexec is a very heated topic. There is no particular technical
rationale for this directory, because a subdirectory under /usr/lib is
generally equivalent. So we shouldn't *recommend* the use of
/usr/libexec. Nevertheless, it is fairly widely used by many
distributions (Fedora/RHEL, openSUSE/SLE, Debian/Ubuntu, Slackware,
OpenEmbedded/Yocto, and Gentoo), so we want to describe this directory
so people know how it is used when they see it on their systems. It is
now described as "secondary" and "alternative". I didn't want to use
the word "legacy", because there is no plan to get rid of it in the
foreseeable future.

Altenative to https://github.com/uapi-group/specifications/pull/149.