LFSH: clarify that systemd tools are not always available
In https://lwn.net/Articles/1032947/#Comments the topic of whether the
mention of systemd-path indicates that it is the normative way to
query this path was heavily contested. The mention of the tool was
originally added as a hint, since systemd-path is only optional. Add
a few words to indicate that systemd-path is optional. (… when first
mentioned. I don't think we want to repeat this every time.)

Similarly, systemd-udev or something else might be used to populate
/dev, so reword as suggested in the LWN comments.