BLS: say that initrd may be split at whitespace for compat
A path is not allowed to contain whitespace (only ASCII upper and lower case
characters, digits, "+", "-", "_" and "." per the first section), so we know
that a space cannot occur in the value. But unfortunately, this is an easy
mistake to make. We made the situation worse by allowing other keys to contain
whitespace-separated paths, e.g. 'devicetree-overlay' is defined as "multiple
overlays are separated by spaces and applied in the same order as they are
listed."  So it's easy to try to apply the same pattern in other places.
Also, the implementation that was done for grub2 allowed multiple paths to be
specified. It is certainly confusing if we say that
  initrd /some/path/initrd /some/path/extra
is not a valid. If instead we were to correctly understand that the user meant
two different initrds and emit a warning, our behaviour would be more robust
and user-friendly.

C.f. https://github.com/systemd/systemd/issues/38834
https://bugzilla.redhat.com/show_bug.cgi?id=2323348
https://lists.fedoraproject.org/archives/list/devel@lists.fedoraproject.org/message/AYB4VHQ5K6CEV4NYCK2QTZB2OEAIJMP7/