bls: tighten rules on file system choice
Let's just say ESP/XBOOTLDR has to be VFAT. Everything else is just
pain, because it means we cannot share the dirs between OSes, systems
and so on.

Moreover, it's a security issue to use more complex file systems, as
these file systems come without integrity protection, i.e. they cannot
be authenticated before parsing them, which makes it key to limit use to
VFAT.

In systemd for example we'll never mount ESP/XBOOTLDR as anything else
than VFAT for these reasons, and the spec should be clear about this.
(well, once https://github.com/systemd/systemd/pull/39267 is merged that is)