specs/lfsh: describe /usr/lib and /usr/lib64 more fully
The spec so far described the multiarch scheme used in Debian and derivatives.
This scheme is great, but for various reasons it hasn't seen wide adoption,
and this means that the spec describes the layout that most users will
actually see only very briefly. The spec also implied that the rules for
multiarch and multilib systems are the same, just with sligtly different
paths. But this is not entirely true, since on multilib systems /usr/lib plays
a similar role as directories one level lower (i.e. /usr/lib/<arch-id>) on
multiarch systems.

Since it's useful to describe the layouts that are commonly used, add a new
section called "Libraries, system programs, and program assets" and describe
/usr/lib, /usr/lib/<arch-id>, /usr/lib64, /usr/libexec there. Explicitly say
that multiarch is the recommended approach, but others are commonly used too.
I think this makes it easier for users to understand that this zoo of
locations is a bikeshedding exercise to a large extent.

Also, change /usr/lib/arch-id to /usr/lib/<arch-id> everywhere.

The description of /usr/share is adjusted a bit. Previously, we said that this
is is the location for data shared between programs and is subject to
interoperability requirements, but quite obviously, a large part of /usr/lib
plays the same role, with data that is shared between programs and subject to
interoperability requirements. So copy the text about interoperability to the
new section too, and say that /usr/share is for architecture-independent data.
(This is the historical origin of "share" in the name.)

The general goal of this change is to make the spec more suitable for multilib
systems.