hugo: use menu element urls relative to the base url
By using relative URLs, which are resolved from the defined base url,
the menu link will not be considered a remote link and will subsequently
leave the link to be opened in the same page. The hugo resolved URL on
build will be the same as the absolute variant.

This helps to reduce future incompatibility by minimizing overwritten
templates.