tpm-pcr-registry: fix design for dark mode
The text in the first two columns of the table was unreadable when using
dark mode. It now stays black, instead of turning white.

Currently this is the exact same color as in light mode. However, if the
underlying theme changes it later on, it needs to be changed manually.
As the color css variable is overwritten in dark mode, I don't think
there is a way to avoid this.

I had a look at the other pages, and it seems like this kind of design
is only used on this page. Therefore I don't think there is any benefit
from moving these styles to an asset css file, especially since it would
make adding additional "Used by" categories a little bit harder.