specs: import OSC 3008 spec from systemd
This stuff is really generic, and fits the purpose of the UAPI group,
hence let's migrate this here.

This is mostly a 1:1 copy of the spec from systemd, but I enclosed the
final part that carries a list of OSC sequences currently known in
markdown comments, because it's not really part of the spec, but still
kinda useful for spec writers.

(Maybe one day we should turn that table into a "registry" style spec of
its own?)